requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.9
openai>=1.60.0
python-frontmatter>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
